+++ pkg/utils/kubernetes/resources/services.go
@@ -159,6 +159,7 @@ func ServicePortsFromDataPlaneIngressOpt(dataplane *operatorv1beta1.DataPlane) S
 					Protocol:   corev1.ProtocolTCP,
 					Port:       p.Port,
 					TargetPort: targetPort,
+					NodePort:   p.NodePort,
 				})
 				alreadyUsedPorts[p.Port] = struct{}{}
 			}